MC-42791: [Magento Cloud] Products not showing in catagories
